fix: resolve unnecessary reloading on mode toggle
