+++ client/src/app/components/HeaderProps.ts
@@ -65,14 +65,4 @@ export interface HeaderProps {
          */
         url: string;
     }[];
-
-    /**
-     * Darkmode theme
-     */
-    darkMode?: boolean;
-
-    /**
-     * Function to toggle dark/light theme
-     */
-    toggleMode?: () => void;
 }
